requirements-dev.txt
coverage
paramiko
pytest
PyYAML
scp
